Dave Perry:
This patch adds main gear rotations about the fuselage attach points that keep the wheels in contact with the ground as the compression increases and decreases.  I also rotated the main fairings and wheels 4 degrees so the wheel axles are horizontal sitting on the ground.  The left and right main gear struts, wheels, and fairings were adjusted to be mirrored (symetric) in the c172p.ac.  I also increased the rpm boundary between the Propeller and Propeller.slow selects.