Initial version of JSBSim hydrodynamics-based floats.
The configuration is based on that of MTB_20m with many adjustments.
The static buoyancy parts have been generated from the floats 3d model.
The various lift and drag coefficients have been tuned for reasonable
behaviour but has not been cross-checked against real performance data.

The hydrodynamics system is not yet completely integrated with the damage
and effects systems.

This commit contains development done from 2015-05-23 to 2015-06-11.

Signed-off-by: Anders Gidenstam <anders@gidenstam.org>