Initial version of JSBSim hydrodynamics-based floats.
The configuration is based on that of MTB_20m and is untuned.
Only the static buoyancy parts have been generated from the floats 3d model.
It appears that either the position of the floats or the CG is wrong as the
aircraft floats nose down. It is probably impossible to achieve take-off
speed currently.