Battery master needs to be on to check fuel gauges
This is not 100% accurate, as per the C172 operating handbook the fuel gauges are checked during the preflight check. However the complete preflight check is not included in the checklists, and the fuel gauge check is instead part of the "before starting engine" checklist, which makes sense in practice. So we need to move the battery master item a few lines higher to be able to see the fuel gauge values. 

In the handbook the battery is turned off afterwards to save the battery since the preflight takes a while, but turning it off here does not make sense, as you'll be starting the engine in a few moments anyway in the sim when you follow the list.