Point to correct gauge
This was pointing at the oil temperature gauge, which is not what you need to look at after the engine starts, it takes longer to move towards green. Fixing to point the marker at oil pressure gauge and scaled it a bit larger as well so you can see the value without the marker blocking it.