Use new ALS light effect for navigation lights
Basic renderer and Rembrandt are still supported. The position of the green
and red navigation lights are translated if the corresponding wing is
damaged. The navlight_{left,right,back} objects, used by the non-ALS
renderers, make use of this animation as well so that all the other
navlight_* objects can be deleted from the Blender files.

Signed-off-by: onox <denkpadje@gmail.com>