C172sp / FG1000: ESS/AVN Bus wiring corrected
According to the POH, the PFD is powered by both, AVN1 and ESS bus.
This is for failsafe reasons.
This commit fixes a bug where the power to the PFD is solely calculated by the ESS breaker;
but the unit should remain on when ESS fails but power is available trough thze avionics bus.