KMA20TSO Audio Panel: made AUTO mode work and make panel failable
According to the handbook, the MIC selector knob also can switch
the audible COM radio, when the AUTO switch is active and the switch
for the radio is in OFF position.
As I interpret this, setting a radio to SPKR or HEADPHONE will make
that radio alwas audible (so its an override for the AUTO switch).

Also, the code now does calculate the "operable" property to make it
failable by setting "serviceable" to false.
The panel has a failsafe mode: when the panel fails, COM1 is
hardwired to the speaker/headset and mic.

see #1527