Commented yoke feedback of aileon trim
Currently no aileron trim feedback to yoke. The C172 has an asymetrically mounted wing in order to
compensate for roll-forces in cruise. This is simulated by the aileron-trim prop.
Aileron _input_ (yoke) should be neutral, when the input device (joystick) is neutral.
Also, don't confuse with rudder-trim (which there is an option for).

For disuccsion see: https://github.com/c172p-team/c172p/issues/1471